recommender_service/requirements.txt
# Основные зависимости
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
implicit==0.6.2
pymongo==4.6.1
redis==5.0.1
python-dotenv==1.0.0
SQLAlchemy==2.0.25
mysqlclient==2.2.1
pymysql==1.1.0
celery==5.3.6
rectools==0.4.0
pydantic-settings==2.7.1
fastapi==0.115.7
uvicorn==0.34.0
threadpoolctl==3.5.0
APScheduler==3.11.0

# Инструменты разработки
pytest==7.4.4
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.10.0